Version 3 refactor
Require Node.js version 8 or higher.
Add npm 5 package-lock file.
Drop babel, the features being used are now supported directly.
Upgradle postcss to version 6.
Update ESLint to version 4.
