--- .babelrc
@@ -1,9 +1,0 @@
-{
-  "comments": false,
-  "compact": true,
-  "plugins": [
-    ["transform-es2015-modules-commonjs-simple", {
-        "addExports": true
-      }]
-  ]
-}
