+++ .gitignore
@@ -1,7 +1,3 @@
-### Custom ###
-dist
-_test-factory.js
-
 ### Node ###
 # Logs
 logs
