+++ package.json
@@ -2,15 +2,12 @@
   "name": "postcss-combine-duplicated-selectors",
   "version": "2.3.0",
   "description": "automatically keep css selectors unique",
-  "main": "dist/index.js",
+  "main": "src/index.js",
   "files": [
     "dist"
   ],
   "scripts": {
-    "build": "babel src -d dist --no-comments",
-    "prepublish": "rimraf dist && npm run build",
-    "pretest": "babel test/_test-factory.es -o test/_test-factory.js --no-comments",
-    "test": "npm run build && ava && eslint --ext md,js,es . && remark *.md -q --no-stdout"
+    "test": "ava && eslint --ext md,js . && remark *.md -q --no-stdout"
   },
   "keywords": [
     "postcss-plugin",
@@ -30,77 +27,46 @@
   },
   "license": "MIT",
   "dependencies": {
-    "postcss": "^5.2.6",
-    "postcss-selector-parser": "^2.2.2"
+    "postcss": "^6.0.1",
+    "postcss-selector-parser": "^2.2.3"
   },
   "devDependencies": {
     "ava": "^0.19.1",
-    "babel-cli": "^6.24.1",
-    "babel-plugin-transform-es2015-modules-commonjs-simple": "^6.7.4",
-    "eslint": "^3.19.0",
+    "eslint": "^4.0.0-beta.0",
     "eslint-config-google": "^0.7.1",
     "eslint-plugin-ava": "^4.2.0",
-    "eslint-plugin-import": "^2.2.0",
     "eslint-plugin-markdown": "^1.0.0-beta.6",
-    "eslint-plugin-node": "^4.2.2",
-    "postcss-less": "^0.15.0",
-    "postcss-nested": "^1.0.1",
-    "postcss-scss": "^0.4.1",
-    "remark-cli": "^3.0.0",
+    "postcss-less": "^1.0.0",
+    "postcss-nested": "^2.0.2",
+    "postcss-scss": "^1.0.0",
+    "remark-cli": "^3.0.1",
     "remark-preset-lint-consistent": "^2.0.0",
     "remark-preset-lint-recommended": "^2.0.0",
-    "remark-validate-links": "^6.0.0",
-    "rimraf": "^2.6.1"
+    "remark-validate-links": "^6.0.0"
   },
   "engines": {
-    "node": ">=4"
+    "node": ">=8"
   },
   "eslintConfig": {
     "root": true,
     "parserOptions": {
-      "ecmaVersion": 8,
-      "sourceType": "module"
+      "ecmaVersion": 8
+    },
+    "env": {
+      "es6": true,
+      "node": true
     },
     "plugins": [
-      "import",
-      "markdown",
-      "node"
+      "markdown"
     ],
     "extends": [
-      "google",
-      "plugin:import/errors",
-      "plugin:import/warnings",
-      "plugin:node/recommended"
+      "eslint:recommended",
+      "google"
     ],
     "rules": {
       "prefer-arrow-callback": "error",
       "prefer-const": "error",
-      "prefer-template": "error",
-      "node/no-unsupported-features": [
-        "error",
-        {
-          "version": 4,
-          "ignores": [
-            "modules"
-          ]
-        }
-      ],
-      "node/no-missing-require": [
-        "error",
-        {
-          "allowModules": [
-            "postcss-combine-duplicated-selectors"
-          ]
-        }
-      ],
-      "node/no-unpublished-require": [
-        "error",
-        {
-          "allowModules": [
-            "postcss-combine-duplicated-selectors"
-          ]
-        }
-      ]
+      "prefer-template": "error"
     }
   },
   "remarkConfig": {
