+++ test/.eslintrc.yml
@@ -2,8 +2,3 @@ extends:
   - plugin:ava/recommended
 plugins:
   - ava
-settings:
-  import/ignore:
-    - node_modules
-    - dist
-    - _test-factory
